LayoutTests/svg/animations/svglength-animation-invalid-value-2-expected.txt

SVG 1.1 dynamic animation tests

Test SVGLength animation with invalid value: No spaces before value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 0
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 0
PASS rect.width.baseVal.value is 100
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
